Add course content tab to instructor dashboard overview
